backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 文件处理
aiofiles==23.2.1
python-magic==0.4.27

# 工具
httpx==0.25.1

# WebSocket
python-socketio==5.10.0
